latest
Modifications : in case of branches, the stems r visited twice as both the 
branch call the func for stem. Due to this the entire path from stem to PI
is covered twice. In case of multiple branches, this path will be visited
multiple times
sol-maintain a list to check whether the node is already visited or not
MODIFICATION done